requirements.txt
streamlit
pandas
scikit-learn
xgboost
fpdf
plotly
Pillow
